feat-strip-story - update Apester Demo App - update ApesterKit Tests
