+++ Apester/Views/APEUIWebViewTableViewCell.swift
@@ -20,7 +20,7 @@ class APEUIWebViewTableViewCell: APEWebViewTableViewCell {
     return webView
   }()
 
-  override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+  override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
     super.init(style: style, reuseIdentifier: reuseIdentifier)
     setupWebContentView(webView: webView)
   }
