+++ Apester/Views/APEWKWebViewTableViewCell.swift
@@ -21,7 +21,7 @@ class APEWKWebViewTableViewCell: APEWebViewTableViewCell {
     return webView
   }()
 
-  override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+  override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
     super.init(style: style, reuseIdentifier: reuseIdentifier)
     setupWebContentView(webView: webView)
   }
