+++ Apester/Views/APEWebViewTableViewCell.swift
@@ -22,7 +22,7 @@ class APEWebViewTableViewCell: UITableViewCell {
   private let initialHeight: CGFloat = 400
   private var heightConstraint: NSLayoutConstraint?
 
-  override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+  override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
     super.init(style: style, reuseIdentifier: reuseIdentifier)
   }
